+++ relay-event-schema/src/protocol/contexts/mod.rs
@@ -19,6 +19,7 @@ mod runtime;
 mod spring;
 mod threadpool_info;
 mod trace;
+mod unity;
 mod user_report_v2;
 pub use app::*;
 pub use browser::*;
@@ -40,6 +41,7 @@ pub use runtime::*;
 pub use spring::*;
 pub use threadpool_info::*;
 pub use trace::*;
+pub use unity::*;
 pub use user_report_v2::*;
 
 use relay_protocol::{Annotated, Empty, FromValue, IntoValue, Object, Value};
@@ -107,6 +109,8 @@ pub enum Context {
     OTAUpdates(Box<OTAUpdatesContext>),
     /// Chromium Stability Report from minidump.
     ChromiumStabilityReport(Box<StabilityReportContext>),
+    /// Unity information.
+    Unity(Box<UnityContext>),
     /// Additional arbitrary fields for forwards compatibility.
     #[metastructure(fallback_variant)]
     Other(#[metastructure(pii = "true")] Object<Value>),